refactor: 북마크 폴더 구조 변경
